moved /state to /core . added react and react dom as devDependencies
