+++ src/core/AlertProvider.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
-import { reducer } from "../state/reducer";
-import { initialState } from "../state/initialState";
+import { reducer } from "./state/reducer";
+import { initialState } from "./state/initialState";
 import { AlertRenderer } from "./AlertRenderer";
-import { Store } from "../state/Store";
-import { remove, addDefaultProps } from "../state/actionCreators";
+import { Store } from "./state/Store";
+import { remove, addDefaultProps } from "./state/actionCreators";
 import { positionsArray } from "./constants";
 
 const reducers = {
