+++ src/core/constants.js
@@ -87,7 +87,7 @@ export const transitionConstants = {
 };
 
 const content =
-  "content: tent: define some con tent: define some con tent: define some con tent: define some condefine some content";
+  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.";
 
 export const defaultAlertArgs = {
   type: info,
