+++ tools/jestSetup.js
@@ -1,10 +1,8 @@
 import Enzyme, { shallow, render, mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import sinon from "sinon";
 
 Enzyme.configure({ adapter: new Adapter() });
 
 global.shallow = shallow;
 global.render = render;
 global.mount = mount;
-global.sinon = sinon;
